requirements.txt
Django==3.0.6
gunicorn==20.0.4
redis==3.5.3
djangorestframework==3.12.2